update wasmtime 25->26

+++ native/wasmex/src/store.rs
@@ -39,7 +39,7 @@ pub struct ExWasiOptions {
 #[module = "Wasmex.StoreLimits"]
 pub struct ExStoreLimits {
     memory_size: Option<usize>,
-    table_elements: Option<u32>,
+    table_elements: Option<usize>,
     instances: Option<usize>,
     tables: Option<usize>,
     memories: Option<usize>,
